Add opt-in minimal permissions run mode
Per some internal discussion, we would rather that the exporter not take
on the responsibility of setting Logpull log retention on zones, but
rather to have this be done via Terraform.

On the other hand, I want this exporter to be as plug-and-play as
possible in the event that we should open source it.

As a compromise, I've added an `EXPORTER_MINIMAL_PERMISSIONS`
environment variable which, when set, disables some initialization
checks and charges forward assuming that the Cloudflare zones already
have log retention enabled.