Remove automagic Cloudflare zone configuration
I believe an exporter is viewed as a read-only tool, and thus the
automatic change of Cloudflare zone configuration is likely surprising
to users. It also requires that the credentials used have `logs:edit`
permissions versus the more constrained `logs:read`.

This configuration is still necessary for the exporter to work, so we
enhance the documentation to include two different ways in which this
may be handled externally: via cURL (per the official documentation)
and via a Terraform `cloudflare_logpull_retention` resource.