Create a logpullAPI client type
The API client provided by cloudflare-go does not currently provide
methods for interacting with Logpull endpoints, beyond to check and
enable log retention.

We have to implement our own logic to interact with this endpoint.
Previously, we had a single function which took a reference to an
existing cloudflare-go API client and plucked some exported properties
off of it in order to construct requests for this endpoint.

This approach works, but it has issues:

1. The only reason this function took a reference to a cloudflare-go API
   client was to read some authentication properties from it in order to
   build its own request. Constructing our own request is necessary, but
   as a user, I'd expect that a function taking an API client would use
   said client to make API requests.

2. There was no good way to mock the Cloudflare API server for writing
   unit tests. We could have passed in our own HTTP client in addition
   to the cloudflare-go client we were already taking, but it makes for
   an ugly API.

To fix this, we create a type: logpullAPI. This type is used like a
special-purpose API client, filling the current gaps in the
cloudflare-go client. It provides three constructors which mimic the API
of the cloudflare-go API client constructors, and is designed do be
easily replaced by cloudflare-go's client if it can provide this
functionality in the future.

Additionally, we define a setAPIProperties method for overriding the
base URL and HTTP client used to issue API requests. This allows for
simpler integration testing with the standard library's httptest
package.